refactor: move formatGQLError util to error utils
