+++ src/utils/common.utils.ts
@@ -1,7 +1,6 @@
 import createCache from "@emotion/cache";
 import axios from "axios";
 import createAuthRefreshInterceptor from "axios-auth-refresh";
-import { SourceLocation } from "graphql";
 import { t } from "i18next";
 import Router from "next/router";
 import React, { isValidElement, ReactNode } from "react";
@@ -83,7 +82,7 @@ export const scrollTop = () => {
   animateScroll.scrollToTop(options);
 };
 
-// TODO: Consider removing redirectTo
+// TODO: Remove redirectTo - unneeded abstraction
 export const redirectTo = (path: string) => Router.push(path);
 
 export const createEmotionCache = () => createCache({ key: "css" });
@@ -97,13 +96,4 @@ export const initAxe = () => {
     const axe = require("@axe-core/react");
     axe(React, ReactDOM, 1000);
   }
-};
-
-export const formatGQLError = (
-  message: string,
-  path?: readonly (string | number)[],
-  locations?: readonly SourceLocation[]
-) => {
-  const locationsStr = JSON.stringify(locations);
-  return `[GraphQL error]: Message: ${message}, Locations: ${locationsStr}, Path: ${path}`;
 };
